backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
boto3==1.34.84
openpyxl==3.1.2
reportlab==4.1.0
Pillow==10.3.0
celery==5.3.6
redis==5.0.4
drf-spectacular==0.27.2
django-simple-history==3.7.0
anthropic==0.40.0
